// src/app/types/TechStack.ts
export type TechStack = {
    key: string; // Unique identifier used to fetch the corresponding icon
    title: string; // Display name of the technology
    href: string; // Official website URL of the technology
    categories: string[];
    theme?: boolean; // If `true`, the icon changes based on dark and light mode
    image?: string;
};